# assets/generated/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cluster Manager API SSH provider",
    "version": "v0.1.0",
    "contact": {
      "name": "Samsung CNCT",
      "url": "https://github.com/samsung-cnct/cma-ssh",
      "email": "samsung.cloudnative@gmail.com"
    }
  },
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/cluster": {
      "get": {
        "summary": "Will retrieve the status of a cluster and its kubeconfig for connectivity",
        "operationId": "GetCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshGetClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "Name of the cluster to be looked up.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "delete": {
        "summary": "Will delete a cluster",
        "operationId": "DeleteCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshDeleteClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "What is the cluster's name to destroy.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "post": {
        "summary": "Will provision a cluster",
        "operationId": "CreateCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshCreateClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmasshCreateClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
...
    "/api/v1/cluster/pool": {
      "delete": {
        "summary": "Will delete a node pool from a provisioned cluster",
        "operationId": "DeleteNodePool",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshDeleteNodePoolReply"
            }
          }
        },
        "parameters": [
          {
            "name": "clusterName",
            "description": "What is the cluster to delete node pools.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "node_pool_names",
            "description": "What is the node pool names to delete.",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "post": {
        "summary": "Will add node pool to a provisioned cluster",
        "operationId": "AddNodePool",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshAddNodePoolReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmasshAddNodePoolMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will scale the number of machines in a node pool for a provisioned cluster",
        "operationId": "ScaleNodePool",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshScaleNodePoolReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmasshScaleNodePoolMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
    "/api/v1/cluster/upgrade": {
      "get": {
        "summary": "Will return upgrade options for a given cluster",
        "operationId": "GetUpgradeClusterInformation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshGetUpgradeClusterInformationReply"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "description": "What is the cluster that we are considering for upgrade.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Cluster"
        ]
      },
      "put": {
        "summary": "Will attempt to upgrade a cluster",
        "operationId": "UpgradeCluster",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/cmasshUpgradeClusterReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/cmasshUpgradeClusterMsg"
            }
          }
        ],
        "tags": [
          "Cluster"
        ]
      }
    },
...
  },
  "definitions": {
...
    "cmasshAddNodePoolMsg": {
      "type": "object",
      "properties": {
        "clusterName": {
          "type": "string",
          "title": "What is the cluster to add node pools to"
        },
        "worker_node_pools": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmasshMachineSpec"
          },
          "title": "What Machines to add to the cluster"
        }
      }
    },
    "cmasshAddNodePoolReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not the node pool was provisioned by this request"
        }
      }
    },
    "cmasshClusterDetailItem": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster"
        },
        "status_message": {
          "type": "string",
          "title": "Additional information about the status of the cluster"
        },
        "kubeconfig": {
          "type": "string",
          "title": "What is the kubeconfig to connect to the cluster"
        },
        "status": {
          "$ref": "#/definitions/cmasshClusterStatus",
          "title": "The status of the cluster"
        }
      }
    },
    "cmasshClusterItem": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster"
        },
        "status_message": {
          "type": "string",
          "title": "Additional information about the status of the cluster"
        },
        "status": {
          "$ref": "#/definitions/cmasshClusterStatus",
          "title": "The status of the cluster"
        }
      }
    },
    "cmasshClusterStatus": {
      "type": "string",
      "enum": [
        "STATUS_UNSPECIFIED",
        "PROVISIONING",
        "RUNNING",
        "RECONCILING",
        "STOPPING",
        "ERROR",
        "DEGRADED"
      ],
      "default": "STATUS_UNSPECIFIED",
      "title": "- STATUS_UNSPECIFIED: Not set\n - PROVISIONING: The PROVISIONING state indicates the cluster is being created.\n - RUNNING: The RUNNING state indicates the cluster has been created and is fully usable.\n - RECONCILING: The RECONCILING state indicates that some work is actively being done on the cluster, such as upgrading the master or node software.\n - STOPPING: The STOPPING state indicates the cluster is being deleted\n - ERROR: The ERROR state indicates the cluster may be unusable\n - DEGRADED: The DEGRADED state indicates the cluster requires user action to restore full functionality"
    },
    "cmasshControlPlaneMachineSpec": {
      "type": "object",
      "properties": {
        "labels": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmasshKubernetesLabel"
          },
          "title": "The labels for the control plane machines"
        },
        "instanceType": {
          "type": "string",
          "title": "Type of machines to provision (standard or gpu)"
        },
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "The number of machines"
        }
      },
      "title": "The specification for a set of control plane machines"
    },
    "cmasshCreateClusterMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "Name of the cluster to be provisioned"
        },
        "k8s_version": {
          "type": "string",
          "description": "The version of Kubernetes for worker nodes. Control plane versions are determined by the MachineSpec."
        },
        "control_plane_nodes": {
          "$ref": "#/definitions/cmasshControlPlaneMachineSpec",
          "title": "Machines which comprise the cluster control plane"
        },
        "worker_node_pools": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmasshMachineSpec"
          },
          "title": "Machines which comprise the cluster"
        }
      }
    },
    "cmasshCreateClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Whether or not the cluster was provisioned by this request"
        },
        "cluster": {
          "$ref": "#/definitions/cmasshClusterItem",
          "title": "The details of the cluster request response"
        }
      }
    },
    "cmasshDeleteClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Could the cluster be destroyed"
        },
        "status": {
          "type": "string",
          "title": "Status of the request"
        }
      }
    },
    "cmasshDeleteNodePoolReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Was this a successful request"
        }
      }
    },
...
    "cmasshGetClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Is the cluster in the system"
        },
        "cluster": {
          "$ref": "#/definitions/cmasshClusterDetailItem"
        }
      }
    },
    "cmasshGetUpgradeClusterInformationReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Can the cluster be upgraded"
        },
        "versions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "What versions are possible right now"
        }
      }
    },
...
    "cmasshKubernetesLabel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "The name of a label"
        },
        "value": {
          "type": "string",
          "title": "The value of a label"
        }
      }
    },
    "cmasshMachineSpec": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "The name of the machine set"
        },
        "labels": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmasshKubernetesLabel"
          },
          "title": "The labels for the machine set"
        },
        "instanceType": {
          "type": "string",
          "title": "Type of machines to provision (standard or gpu)"
        },
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "The number of machines"
        }
      },
      "title": "The specification for a set of machines"
    },
    "cmasshScaleNodePoolMsg": {
      "type": "object",
      "properties": {
        "clusterName": {
          "type": "string",
          "title": "What is the name of the cluster to scale a node pool"
        },
        "node_pools": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/cmasshScaleNodePoolSpec"
          },
          "title": "What node pools to scale"
        }
      }
    },
    "cmasshScaleNodePoolReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Was this a successful request"
        }
      }
    },
    "cmasshScaleNodePoolSpec": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "What is the node pool name to scale"
        },
        "count": {
          "type": "integer",
          "format": "int32",
          "title": "Number of machines to scale"
        }
      }
    },
    "cmasshUpgradeClusterMsg": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "What is the cluster that we are considering for upgrade"
        },
        "version": {
          "type": "string",
          "title": "What version are we upgrading to?"
        }
      }
    },
    "cmasshUpgradeClusterReply": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean",
          "format": "boolean",
          "title": "Was this a successful request"
        }
      }
    }
  },
...
}